separated some functions
